use float64 for precision

+++ pkg/models/client_lot.go
@@ -13,10 +13,10 @@ type ClientLotModel struct {
 	ClientID    primitive.ObjectID  `bson:"clientID" json:"clientID"`
 	AgentID     *primitive.ObjectID `bson:"agentID" json:"agentID"`
 	Status      string              `bson:"status" json:"status"`
-	Price       float32             `bson:"price" json:"price" validate:"number,min=0"`
-	DownPayment float32             `bson:"downPayment" json:"downPayment" validate:"required,number,min=0"`
+	Price       float64             `bson:"price" json:"price" validate:"number,min=0"`
+	DownPayment float64             `bson:"downPayment" json:"downPayment" validate:"required,number,min=0"`
 	Months      int                 `bson:"months" json:"months" validate:"required,number,min=0"`
-	Monthly     float32             `bson:"monthly" json:"monthly" validate:"required,number,min=0"`
+	Monthly     float64             `bson:"monthly" json:"monthly" validate:"required,number,min=0"`
 	Date        time.Time           `bson:"date" json:"date" validate:"required"`
 	ApprovedBy  *primitive.ObjectID `bson:"approvedBy" json:"approvedBy"`
 	// Extended
