+++ pkg/models/lot.go
@@ -6,9 +6,9 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type LotModel struct {
 	ProjectID  *primitive.ObjectID `bson:"projectID" json:"projectID"`
 	Name       string              `bson:"name" json:"name" validate:"required"`
-	Area       float32             `bson:"area" json:"area" validate:"number,min=0"`
-	Price      float32             `bson:"price" json:"price" validate:"number,min=0"`
-	PriceAddon float32             `bson:"priceAddon" json:"priceAddon" validate:"number,min=0"`
+	Area       float64             `bson:"area" json:"area" validate:"number,min=0"`
+	Price      float64             `bson:"price" json:"price" validate:"number,min=0"`
+	PriceAddon float64             `bson:"priceAddon" json:"priceAddon" validate:"number,min=0"`
 	Status     string              `bson:"status" json:"status"`
 	// Extended
 	BaseModel `bson:",inline"`
